requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-groq
python-dotenv
pandas
plotly
scikit-learn
langchain-huggingface
joblib